+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - Additional console commands for handling the BroadcastManager
+
+### Changed
+ - Fixed ObjectDisposedException in Net Standard when stopping the AsyncUdpClient
 
 ## [4.0.0] - 2018-04-15
 ### Added