+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
  - Adding an explicit BroadcastData JSON converter
+ - Additional console commands for handling the BroadcastManager
+
+### Changed
+ - Fixed ObjectDisposedException in Net Standard when stopping the AsyncUdpClient
 
 ## [4.0.0] - 2018-04-15
 ### Added